Reworked CameraController to CameraPositionLimiter
- When moving the CAVEPlayer / XROrigin, the "Cameras" GameObject would get stuck behind as the CameraController would limit its global position to the bounds
- CameraPositionLimiter got all the deprecated movement code of CameraController removed and is working with localPosition instead of globalPosition
